pwm: Convert period and duty cycle to u64
With the changes from a9d887dc, the pwm_capture members were left
unmodified. This commit adds the missing changes.

+++ include/linux/pwm.h
@@ -327,8 +327,8 @@ struct pwm_chip {
  * @duty_cycle: duty cycle of the PWM signal (in nanoseconds)
  */
 struct pwm_capture {
-	unsigned int period;
-	unsigned int duty_cycle;
+	u64 period;
+	u64 duty_cycle;
 	unsigned int phase;
 };
 
@@ -347,8 +347,8 @@ int pwm_adjust_config(struct pwm_device *pwm);
  *
  * Returns: 0 on success or a negative error code on failure.
  */
-static inline int pwm_config(struct pwm_device *pwm, int duty_ns,
-			     int period_ns)
+static inline int pwm_config(struct pwm_device *pwm, u64 duty_ns,
+			     u64 period_ns)
 {
 	struct pwm_state state;
 
@@ -457,8 +457,8 @@ static inline int pwm_adjust_config(struct pwm_device *pwm)
 	return -ENOTSUPP;
 }
 
-static inline int pwm_config(struct pwm_device *pwm, int duty_ns,
-			     int period_ns)
+static inline int pwm_config(struct pwm_device *pwm, u64 duty_ns,
+			     u64 period_ns)
 {
 	return -EINVAL;
 }
